% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{package}
\name{DigitalPicoTools}
\alias{DigitalPicoTools}
\alias{DigitalPicoTools-package}
\title{DigitalPicoTools : A set of tools  for Digital analysis of whole genome sequencing of picogram quantities of DNA}
\description{
The main functions provided by the tools are : \code{\link{getVariantAlleleInfo}}, \code{\link{getLFRset}} and \code{\link{getPhasingInfo}}
}
\details{
The package include experimental data retrieved from a study conducted in the Ovarian Cancer Lab of the University of Oxford.

For more detailed information on usage, see the package vignette, by typing
\code{vignette('DigitalPicoTools')}. All support questions should be emailed to the authors.
}
\author{
Donatien Chedom-Fotso, Ahmed Ahmed, Christopher Yau,
}
\references{
DigitalPicoTools references:

OncoPhase: A package for computing Somatic Mutation cellular Prevalence in cancer using haplotype phasing. Bioinformatics 2016. Submitted

DigitalPicoTools reference (Paper):

 Ovarian cancer haplotype sequencing reveals ubiquitous SOX2 overexpression in the premalignant fallopian tube epithelium
}
\keyword{package}

