% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getBreadthGenomeCoverage,LFRset-method}
\alias{getBreadthGenomeCoverage,LFRset-method}
\alias{getBreadthGenomeCoverage,getBreadthGenomeCoverage-method}
\title{Compute the percentage of the genome covered by the LFR set.}
\usage{
\S4method{getBreadthGenomeCoverage}{LFRset}(object)
}
\description{
Compute the percentage of the genome covered by the LFR set.
}

