% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{getLFRset}
\alias{getLFRset}
\title{Computing Long fragment Reads from the wells}
\usage{
getLFRset(wells_list, mindistance, unexists.action = "unexists.fail",
  minNbReads = 2, minFragmentsLength = 1000)
}
\arguments{
\item{mindistance}{The minimum distance separating two consecutive LFR. Two paired reads located within that distance are considered to be originating from the same fragments of DNA.}

\item{unexists.action}{Define the action to perform when a well bam file do not exists. Can be one of the following:
\describe{
  \item{\emph{unexists.pass}}{No particular action, consider the provided bam file and continue.}
  \item{\emph{unexists.fail}}{Interupt the run and generate an error}
  \item{\emph{unexists.exclude}}{Do not process the sample, skipt it or remove it from the list of well and continue}
 }
 The default value is set at 'unexists.fail'}

\item{minFragmentsLength}{Minimum length of a fragment for it to be considered as valid. default : 1000}

\item{well_list}{A list of objects of class \code{\link{WellSample}} representing the wells to compute the fragments from. well_list can also be a vector of character string representing the file names to each well BAM file.}

\item{minReads}{Minimum number of reads a fragment should contains to be considered as valid. Default value : 2}
}
\value{
An object of class  \code{\link{LFRset}}.
}
\description{
This function takes as input a list of wells and compute the list of Long Fragments Reads sequenced in each well. Each LFR is represented by its name, its chromosome, its start and end position, its length, its average coverage, the number of reads making up the fragment and the well ID it is originating from. The function return an objoct of class \code{\link{LFRSet}}
}
\note{
''.
}
\examples{
  wellsID_file=system.file('extdata','wells_id_tissue.txt',package='DigitalPicoTools')
  wellsID_list<-unlist(read.table(wellsID_file))
  WellSample_list=initsWellSamples(wellsID_list,
   BAMprefix=paste(path.package('DigitalPicoTools'),  'extdata/',sep='/'))
  FiveWellsLFRset = getLFRset(WellSample_list,3000)
  print(head(FiveWellsLFRset))

  # An object of class LFRdataframe with   rows and   columns
  #                   LFR_name Chrom  Start    End Length AvgCoverage Nbreads Well_Number Well_ID
  # 1   A005_chr1_47518_51209  chr1  47518  51209   3691   0.2915199       4           1    A005
  # 2   A005_chr1_64725_72351  chr1  64725  72351   7626   0.2265932       7           1    A005
  # 3   A005_chr1_77103_78637  chr1  77103  78637   1534   0.6147327       4           1    A005
  # 4   A005_chr1_82636_87332  chr1  82636  87332   4696   0.3017462       6           1    A005
  # 5  A005_chr1_98437_104080  chr1  98437 104080   5643   0.3202197       7           1    A005
  # 6 A005_chr1_110942_113908  chr1 110942 113908   2966   0.2862441       3           1    A005
}
\seealso{
\code{\link{getWellLFRset}}, \code{\link{getLFRStats}, \code{\link{plotLFRDistribution}}}, \code{\link{getWells}},  \code{\link{plotLFRDistribution}}, \code{\link{plot.LFRset}}, \code{\link{summary.LFRset}}, \code{\link{getLFRStatsPerWell}}, \code{\link{getBreadthGenomeCoverage}},  \code{\link{getN50}}
}

