% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getLFRStats,LFRset-method}
\alias{getLFRStats,LFRset-method}
\alias{getLFRStats,getLFRStats-method}
\title{A function returning the following statistics from an LFR dataset.}
\usage{
\S4method{getLFRStats}{LFRset}(object)
}
\description{
A function returning the following statistics from an LFR dataset.
}

