% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getLFRStatsPerWell}
\alias{getLFRStatsPerWell}
\title{A function returning the following statistics per wells from an LFR dataset}
\usage{
getLFRStatsPerWell(object)
}
\arguments{
\item{x}{an object of class \code{\link{LFRset}}}
}
\value{
A data frame  containing for each well having an LFR in the LFR set  the following statistics :N, N50_value,N50_totallength,L50_value,L50_percent,Max,Min,Mean,Median .
}
\description{
\describe{
  \item{N}{The number of Long Fragments reads in the set}
  \item{N50_value}{The N50 of the LFR set}
  \item{N50_totallength}{The Sum of LFR lengths greater or equal to the N50}
  \item{L50}{The L50 of the set of LFR}
  \item{L50_percent}{Percentage of LFR having a length greater or equal to the N50}
  \item{Max}{Maximum LFR length in the set}
  \item{Min}{Minimum LFR length in the set }
  \item{Mean}{TThe mean of the LFR lengths}
  \item{Median}{The median of the LFR length}
}
}
\examples{
data(FiveWellsLFRset)
P009_LFR=getWellLFRset(FiveWellsLFRset,'P009')
print(getLFRStatsPerWell(FiveWellsLFRset))
}
\seealso{
\code{\link{getN50}}, \code{\link{getLFRStats}}, \code{\link{getBreadthGenomeCoverage}}, \code{\link{plotLFRDistribution}}, \code{\link{summary.LFRset}}, \code{\link{getLFRset}}, \code{\link{plot.LFRset}}
}

