% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getLFRStatsPerWell,LFRset-method}
\alias{getLFRStatsPerWell,LFRset-method}
\alias{getLFRStatsPerWell,getLFRStatsPerWell-method}
\title{A function returning the following statistics per wells from an LFR dataset}
\usage{
\S4method{getLFRStatsPerWell}{LFRset}(object)
}
\description{
A function returning the following statistics per wells from an LFR dataset
}

