% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{getMutationsOfLFR}
\alias{getMutationsOfLFR}
\title{Retrieve the LFR's mutations}
\usage{
getMutationsOfLFR(LFR_df, Mutations_df, LFRname, calltype = NULL)
}
\description{
This function computes the list of mutations present in each Long Fragment Reads either on their Variant Allele or their reference Allele
}

