% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getN50}
\alias{getN50}
\title{Compute the N50 value of an LFRset}
\usage{
getN50(object)
}
\arguments{
\item{x}{an object of class \code{\link{LFRset}}}
}
\value{
A numeric containing the N50 value  of the LFR set.
}
\description{
Compute the N50 value of an LFRset
}
\examples{
data(FiveWellsLFRset)
print(getN50(FiveWellsLFRset))
}
\seealso{
\code{\link{getBreadthGenomeCoverage}}, \code{\link{getLFRStats}}, \code{\link{getLFRStatsPerWell}}, \code{\link{LFRset}}, \code{\link{getWellLFRset}},
}

