% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getN50,LFRset-method}
\alias{getN50,LFRset-method}
\alias{getN50,getN50-method}
\title{Compute the N50 of an LFR set.}
\usage{
\S4method{getN50}{LFRset}(object)
}
\description{
Compute the N50 of an LFR set.
}

