% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{getVariantAlleleInfo}
\alias{getVariantAlleleInfo}
\title{Retrieveing variants genomic and alleles count information}
\usage{
getVariantAlleleInfo(VCFFilePath, wells_id = NULL, region = NULL,
  OnlyPassFilter = FALSE, FiltersToInclude = NULL,
  FiltersToExclude = NULL, RetrieveWellsIDs = FALSE)
}
\arguments{
\item{VCFFilePath}{A character string representing the path to a VCF file obtained from an LFR experiment study. The file should contains one genotype information column per well sample.}

\item{wells_id}{A vector of character string representing the well sample ID to assign to the columns name corresponding to each sample genotyping information. If not provided the column names in the VCF file are maintained.  length(well_id) should be equal to the number of column of the VCF file - 9 following to the VCF format specifications.}

\item{region}{A character string representing the region of the genome to consider given in the following format :  'chrom:start-end'. Default value :NULL and all the variants in the VCF file will be considered.}

\item{OnlyPassFilter}{A logical value. If set to true, only the variant having their calling filter status set at 'PASS'}

\item{FiltersToInclude}{A vector of character string representing a list of calling filter status. All the variant having any of them will be included in the analysis.}

\item{FiltersToExclude}{A vector of character string representing a list of calling filter status. All the variant having any of them will be excluded from  the analysis given that the filte ris not included in FiltersToInclude.}

\item{RetrieveWellsIDs}{A logical value specifying if the wells ID supporting the reference and  the SNP should be retrieved or not. Usefull for the Phasing analysis. (Default : FALSE)}
}
\value{
An object of class  \code{\link{VariantAlleleInfo}}.
}
\description{
This function takes as input a VCF file obtained from an LFR experiment and output and returns of class \code{\link{VariantAlleleInfo}} containing the genomic and allele count information of each variants.
}
\note{
The function can also be used for a multiple sample and non LFR experiment. In this case, the well allele count and fraction  information will assess the presence of the variant across the different samples.
}
\examples{
 vcffile<-system.file('extdata','LFR_11152_Tissue_chr22.vcf.gz',package='DigitalPicoTools')
 AlleleInfo = getVariantAlleleInfo(vcffile)
 print(head(AlleleInfo))
 #                     Chrom      Pos REF ALT Qual          Filter TC TR        AF WR WV        WF     WFAdj
 # chr22_16060515_CTT_C chr22 16060515 CTT   C   68  SC;badReads;MQ  4  3 0.7500000  4  3 0.7500000 0.7500000
 # chr22_16114149_C_A   chr22 16114149   C   A   36  SC;badReads;MQ 10  2 0.2000000 10  2 0.2000000 0.2000000
 # chr22_16114636_CT_C  chr22 16114636  CT   C   28  badReads;MQ;QD 13  4 0.3076923 11  4 0.3636364 0.3636364
 # chr22_16114662_G_A   chr22 16114662   G   A  147     badReads;MQ 13  5 0.3846154  8  5 0.6250000 0.6250000
 # chr22_16194000_A_AT  chr22 16194000   A  AT   39     badReads;MQ  2  2 1.0000000  2  2 1.0000000 0.9558641
 # chr22_16228052_T_C   chr22 16228052   T   C   18 Q20;badReads;MQ  6  2 0.3333333  5  2 0.4000000 0.4000000
}
\seealso{
\code{\link{getVariantCoverageTable}}, \code{\link{plot.VariantAlleleInfo}}, \code{\link{SampleCoverage}}, \code{\link{VariantAlleleInfo}}
}

