% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{getVariantCoverageTable}
\alias{getVariantCoverageTable}
\title{Retrieveing variants genomic and alleles count information per sample}
\usage{
getVariantCoverageTable(VCFFilePath, wells_id = NULL, region = NULL,
  OnlyPassFilter = FALSE, FiltersToInclude = NULL,
  FiltersToExclude = NULL)
}
\arguments{
\item{VCFFilePath}{A character string representing the path to a VCF file obtained from an LFR experiment study. The file should contains one genotype information column per well sample.}

\item{wells_id}{A vector of character string representing the well sample ID to assign to the columns name corresponding to each sample genotyping information. If not provided the column names in the VCF file are maintained.  length(well_id) should be equal to the number of column of the VCF file - 9 following to the VCF format specifications.}

\item{region}{A character string representing the region of the genome to consider given in the following format :  'chrom:start-end'. Default value :NULL and all the variants in the VCF file will be considered.}

\item{OnlyPassFilter}{A logical value. If set to true, only the variant having their calling filter status set at 'PASS'}

\item{FiltersToInclude}{A vector of character string representing a list of calling filter status. All the variant having any of them will be included in the analysis.}

\item{FiltersToExclude}{A vector of character string representing a list of calling filter status. All the variant having any of them will be excluded from  the analysis given that the filte ris not included in FiltersToInclude.}
}
\value{
An object of class  \code{\link{SampleCoverage}}.
}
\description{
This function takes as input a VCF file obtained from an LFR experiment and output and returns of class \code{\link{SampleCoverage}} containing the genomic and allele count information per sample of each variants.
}
\note{
The function can also be used on a multiple sample and non LFR experiment.
}
\examples{
 vcffile<-system.file('extdata','LFR_11152_Tissue_chr22.vcf.gz',package='DigitalPicoTools')
 CoverageTable = getVariantCoverageTable(vcffile)
 print(head(CoverageTable))
}
\seealso{
\code{\link{getVariantAlleleInfo}}, \code{\link{plot.VariantAlleleInfo}}, \code{\link{SampleCoverage}}, \code{\link{VariantAlleleInfo}}
}

