% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{getWellLFRset}
\alias{getWellLFRset}
\alias{getWellLFRset,LFRset-method}
\alias{getWellLFRset,getWellLFRset-method}
\title{Retrieves all the LFRs oiginating from a given well}
\usage{
getWellLFRset(object, well)

\S4method{getWellLFRset}{LFRset}(object, well)
}
\arguments{
\item{well}{either an object of class \code{\link{numeric}} representing the well number or an object of class \code{\link{character}} repreenting the well nanme.}

\item{x}{an object of class \code{\link{LFRset}}}
}
\value{
An object of class \code{\link{LFRset}} containing the Long Fragments Reads in x located in the well \code{well}
}
\description{
Retrieves all the LFRs oiginating from a given well

Retrieves all the LFRs oiginating from a given well
}
\examples{
#Loading an object of class (LFRset)
data(FiveWellsLFRset)
print(getWells(FiveWellsLFRset))
#A005  A016  L006  L024  P009
#43755 15192 23448 21892 24397
well1_LFR=getWellLFRset(FiveWellsLFRset,1)
print(getWells(well1_LFR))
#A005
#43755
P009_LFR=getWellLFRset(FiveWellsLFRset,'P009')
print(getWells(P009_LFR))
#P009
#24397
}
\seealso{
\code{\link{getLFRset}}, \code{\link{getLFRStats}}, \code{\link{getLFRStatsPerWell}}, \code{\link{summary.LFRset}}
}

