% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{hist.LFRset}
\alias{hist.LFRset}
\title{Plotting histograms on an LFR data set}
\usage{
\method{hist}{LFRset}(object, value = "Coverage", minlength = 10000,
  color = "black", main = "", well = NULL)
}
\arguments{
\item{minlength}{Minimum length of a fragment to be considered. Default 10000.}

\item{color}{histogram color. default: black}

\item{main}{The title of the plot. default: None}

\item{x}{an  \code{\link{LFRset}} object containing the list of Long Fragment Reads}

\item{Value}{The variable to plot the histogram for . Can be either 'Coverage' either 'length'. default is 'Coverage'.}
}
\value{
''
}
\description{
This is a generic function to plot  the histogram of  the Length or the Average coverage of the long Frangment Reads within an LFR data set. given the paramter \emph{value}, two values can be plotted: The fragment length and the fragment average coverage
}
\examples{

 #Example 1:
 wellsID_file=system.file('extdata','wells_id_tissue.txt',package='DigitalPicoTools')
 wellsID_list<-unlist(read.table(wellsID_file))
 WellSample_list=initsWellSamples(wellsID_list,
     BAMprefix=paste(path.package('DigitalPicoTools'),'extdata/',sep='/'))
 LFR_Info = getLFRset(WellSample_list,3000)
 well1=getWellLFRset(LFR_Info,1)
 hist(well1,value='Length')
 hist(LFR_Info,value='Coverage', color='magenta',main='test')
}
\seealso{
\code{\link{plotLFRDistribution}}, \code{\link{plot.LFRset}}, \code{\link{getWells}}, \code{\link{getWellLFRset}}, \code{\link{getLFRStats}}
}

