% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{initsWellSamples}
\alias{initsWellSamples}
\title{Initialising a set of Well samples}
\usage{
initsWellSamples(WellSamples_list, BAMprefix = NULL, BAMsuffix = NULL,
  unexists.action = "unexists.fail")
}
\arguments{
\item{unexists.action}{Define the action to perform when a bam file do not exists. Can be one of the following:
\describe{
  \item{\emph{unexists.pass}}{No particular action, consider the provided bam file and continue.}
  \item{\emph{unexists.fail}}{Interupt the run and generate an error}
  \item{\emph{unexists.exclude}}{Do not process the sample, remove it from the list of well and continue}
}
The default value is set at \emph{unexists.fail}}

\item{wellSamples_list}{a vector  of well ID.}

\item{Bamprefix}{a character string representing a prefix to add to the well ID to match the bam file name in the system file. Should contains the directories making path to the file. Putting the absolute path to the file is recommaeded.  default value : empty string.}

\item{Bamsuffix}{a character string representing a suffix to add after the well_ID in order to match the Bam file name on the file system. This suffux should not include the extemsion '.bam'. Defaults value : empty string}
}
\value{
A list of objects of class \code{\link{WellSample}}
}
\description{
This function take as input a vector containing a series of well ID and output of list of \code{\link{WellSample}} objects corresponding to each well.
}
\examples{
wellsID_file=system.file('extdata','wells_id_tissue.txt',package='DigitalPicoTools')
 wellsID_list<-unlist(read.table(wellsID_file))
 WellSample_list=initsWellSamples(wellsID_list,
    BAMprefix=paste(path.package('DigitalPicoTools'),'extdata/',sep='/'))
 print(WellSample_list[3])
 # An S4 object of class WellSample
 #
 # Well_ID :  L006
 # Well_Number :  3
 # BAMfilename :  extdata/L006.bam
 #
}
\seealso{
\code{\link{WellSample}}
}

