% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{class}
\name{LFRdataframe-class}
\alias{LFRdataframe}
\alias{LFRdataframe-class}
\title{The LFRdataframe class}
\description{
An S4 base class. All other LFR class containing a data frame Table as main object inherit from this base class.
}
\section{Slots}{

\describe{
\item{\code{Table}}{Object of class \code{\link{data.frame}}, main container of the \code{\link{LFRdataframe}} class}
}}
\seealso{
\code{\link{LFRset}}, \code{\link{SampleCoverage}}, \code{\link{VariantAlleleInfo}}
}

