% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{class}
\name{LFRset}
\alias{LFRset}
\alias{LFRset-class}
\title{The LFRset class}
\description{
An S4 class containing the set of Long Fragments Reads (LFR) retrieved from an LFRSeq experiment.
}
\section{Slots}{

\describe{
\item{\code{Table}}{A data frame serving as a containers of the LFRs. Each LFR is represented by the following fields :
\describe{
  \item{LFRName}{The name of the LFR. By default it is a combination of the well's name the LFR is originating from and the genomic location of the LFR (Chromosome, atart position and end position)}
  \item{Chrom}{The chromosome of the LFR}
  \item{Start}{The starting position  of the LFR}
  \item{End}{The ending position of the LFR}
  \item{Length}{The length of the LFR}
  \item{AvgCoverage}{The average Coverage of the LFR}
  \item{NbReads}{The number of reads making up the LFR}
  \item{Well_ID}{The name or ID of the well the LFR is originating from}
  \item{Well_Number}{The number of the well}
}}
}}
\note{
This class is one of the main class of the package
}
\seealso{
\code{\link{getLFRset}}, \code{\link{getWells}}, \code{\link{getLFRStats}}, \code{\link{plotLFRDistribution}}, \code{\link{plot.LFRset}}
}

