% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{plot.LFRset}
\alias{plot.LFRset}
\title{Plotting an LFR data set}
\usage{
\method{plot}{LFRset}(object, type = "ScatterPlot", Value = "Coverage",
  minlength = 10000, color = "black", main = "", well = NULL)
}
\arguments{
\item{type}{the type of the plot. Can be either 'ScatterPlot' either 'Landscape'. The default value is 'ScatterPlot'.}

\item{Value}{The variable to represent on the y-Axis. Can be either 'Coverage' either 'length'. default is 'Coverage'.}

\item{minlength}{Minimum length of a fragment to be considered. Default 10000.}

\item{color}{The points color. default: black}

\item{main}{The title of the plot. default: None}

\item{x}{an  \code{\link{LFRset}} object containing the list of Long Fragment Reads}
}
\value{
''
}
\description{
This is a generic function to plot  in different way the Length and the Average coverage of the long Frangment Reads within an LFR data set. Two type of plot can be performed.
\describe{
  \item{ScatterPlot}{this plot represent a scatter plot between the Length and the Average Coverage}
  \item{Landscape}{This plot type represent the distribution of the LFR across the whole genome. It adds a third dimension to the length and the coverage which is the genomic location. }
}
}
\details{
In both types, the paramater \emph{Value} allow to specifie which variable will be represented on the y-axis. It can be  'Coverage' for the LFR Average Coverage and 'Length' for the LFR lengths.
}
\examples{

 #Example 1:
 wellsID_file=system.file('extdata','wells_id_tissue.txt',package='DigitalPicoTools')
 wellsID_list<-unlist(read.table(wellsID_file))
 WellSample_list=initsWellSamples(wellsID_list,
     BAMprefix=paste(path.package('DigitalPicoTools'),'extdata/',sep='/'))
 LFR_Info = getLFRset(WellSample_list,3000)
 well1=getWellLFRset(LFR_Info,1)
 plot(well1,Value='Length')
 plot(LFR_Info,type='Landscape', color='magenta',main='test')
 plot(well1,Value='Coverage',type='Landscape')
}
\seealso{
\code{\link{plotLFRDistribution}}, \code{\link{plot.LFRset}}, \code{\link{getWells}}, \code{\link{getWellLFRset}}, \code{\link{getLFRStats}}
}

