% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{plotLFRDistribution}
\alias{plotLFRDistribution}
\title{Plot the LFR distribution}
\usage{
plotLFRDistribution(object, chrom, region = NULL, minLength = 10000)
}
\arguments{
\item{chrom}{the chromosome to be considered}

\item{region}{The genomic region to plot. For best results consider a region of less than 2MB.}

\item{minLength}{Conside ronly the LFR having a length greater or equal to minLength.}

\item{x}{an object of class \code{\link{LFRset}}}
}
\value{
''
}
\description{
Plot the distribution of the LFR across the genome and across all the wells. Suitable to be call on genomic region of less than 2MB length. Segments represent fragments, the y-axis is the wells and the x-axis the genomic location. The thickness of the segment is proportional to the averagecoverage. The blue segment is an illustration of a fagment of 100kbp (if Region <5Mb) or 1Mb if (region >5Mb)  with and average coverage of 1.
}
\examples{
cat('\\n Loading an LFRset data frame ')
LFR_11152Tissue_10k_file<-system.file('extdata',
       'D1000_10k_allfragments.tsv.gz',package = 'DigitalPicoTools')
D1000_10k_11152Tissue_LFR=read.table(file=LFR_11152Tissue_10k_file)
D1000_10k_11152Tissue_LFRset=LFRset(Table=D1000_10k_11152Tissue_LFR)
cat('\\n Plotting the Fragment Distribution ')
plotLFRDistribution(D1000_10k_11152Tissue_LFRset,'chr2',region='chr22:22000000-23000000')
}
\seealso{
\code{\link{plot}}, code{\link{summary.LFRset}}, \code{\link{getLFRset}}, \code{\link{getWells}}
}

