% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{methods}
\name{plotLFRDistribution,LFRset-method}
\alias{plotLFRDistribution,LFRset-method}
\alias{plotLFRDistribution,plotLFRDistribution-method}
\title{A method for plotting the LFR distribution}
\usage{
\S4method{plotLFRDistribution}{LFRset}(object, region = NULL,
  minLength = 10000)
}
\description{
A method for plotting the LFR distribution
}

