% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{class}
\name{SampleCoverage}
\alias{SampleCoverage}
\alias{SampleCoverage-class}
\title{The SampleCoverage class}
\description{
An S4 class containing the genomic location and allele information per sample for each variant from a VCF file.
}
\section{Slots}{

\describe{
\item{\code{Table}}{A data frame containing the genomic location and allele information per sample for each variant. It gives for each called variants the following information :
\describe{
  \item{Chrom}{The chromosome of the variant)}
  \item{Pos}{The position of the variant}
  \item{REF}{The reference sequence }
  \item{ALT}{The altered sequence}
  \item{Qual}{The quality of the variant call}
  \item{Filter}{The VCF filter of the variant call}
  \item{TC}{The total reads count supporting the variant}
  \item{TR}{The total reads count supporting the variant}
  \item{AF}{The reads allele fraction}
  \item{}{For each  well samples}
  \describe{
     \item{NR_sampleID}{The total numbe rof reads covering the variant location at this smaple}
     \item{NV_sampleID}{Number of reads supporting the variant sequence at this smaple}
     \item{AF_sampleID}{Variant Reads allele fraction at this sample}
  }
  \item{GT}{colon separated list of unphased genotype at each sample}
}}
}}
\seealso{
\code{\link{VariantAlleleInfo}}, \code{\link{getVariantAlleleInfo}}, \code{\link{getVariantCoverageTable}}, \code{\link{plot.VariantAlleleInfo}}
}

