% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\name{summary.LFRset}
\alias{summary.LFRset}
\title{Summary statistics on an LFRset Object}
\usage{
summary.LFRset(object)
}
\arguments{
\item{x}{an  \code{\link{LFRset}} object containing the list of Long Fragment Reads}
}
\value{
The five statistics summary (Min, Max, Mean, Median, 1st Quart, 3rd Quart) for the Length, the Average Coverage and the Number of Reads of the LFRs.
}
\description{
This is a generic function to compute the common summary statistic on the Length,
Average coverage and the numbe rof reads of an LFR set object
}
\examples{
LFR_11152Tissue_10k_file<-system.file('extdata',
      'D1000_10k_allfragments.tsv.gz',package = 'DigitalPicoTools')
D1000_10k_11152Tissue_LFR=read.table(file=LFR_11152Tissue_10k_file)
D1000_10k_11152Tissue_LFRset=LFRset(Table=D1000_10k_11152Tissue_LFR)
summary(D1000_10k_11152Tissue_LFRset)
#  Summary on the LFRset object
#
# Fragments Length
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 10000   12340   15800   18700   21920  150200
#
# Fragments Average Coverage
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 0.2747  0.9621  1.3730  1.7560  2.1050 28.9100
#
# Fragments Nbreads
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
# 12.0    53.0    90.0   154.4   174.0  5240.0

}
\seealso{
\code{\link{getBreadthGenomeCoverage}}, \code{\link{getLFRStats}}, \code{\link{getLFRset}},  \code{\link{getLFRStatsPerWell}}, \code{\link{getN50}}
}

