% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{class}
\name{VariantAlleleInfo}
\alias{VariantAlleleInfo}
\alias{VariantAlleleInfo-class}
\title{The VariantAlleleInfo class}
\description{
An S4 class containing the allele information for each variant from a VCF file.
}
\section{Slots}{

\describe{
\item{\code{Table}}{A data frame containing the Allele information per variant. It gives for each called variants the following information :
\describe{
  \item{Chrom}{The chromosome of the variant)}
  \item{Pos}{The position of the variant}
  \item{REF}{The reference sequence }
  \item{ALT}{The altered sequence}
  \item{Qual}{The quality of the variant call}
  \item{Filter}{The VCF filter of the variant call}
  \item{TC}{The total reads count supporting the variant}
  \item{TR}{The total reads count supporting the variant}
  \item{AF}{The reads allele fraction}
  \item{WR}{The total well count at the position of the variant}
  \item{WV}{The count of well supporting the variant}
  \item{Well_Number}{The wells allele fraction}
}}
}}
\note{
This class is one of the main class of the package
}
\seealso{
\code{\link{SampleCoverage}}, \code{\link{VariantAlleleInfo}}, \code{\link{getVariantAlleleInfo}},
         \code{\link{getVariantCoverageTable}}, \code{\link{plot.VariantAlleleInfo}}
}

