% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DigitalPicoTools.R
\docType{class}
\name{WellSample}
\alias{WellSample}
\alias{WellSample-class}
\title{The WellSample class}
\description{
An S4 class containing the main information (Well ID, well Number, Bam file) of a well sample.
}
\section{Slots}{

\describe{
\item{\code{Well_ID}}{Object of class \code{\link{character}} representing the ID or the name assigned to the well}

\item{\code{Well_Number}}{Object of class \code{\link{numeric}} representing the number of the well}

\item{\code{Bamfilename}}{Object of class \code{\link{character}} representing the file name of the well bam file.}
}}
\examples{
 #Creating an instance of the class
 myWell = WellSample(Well_ID='W384',Well_Number=384,BAMfilename='extdata/P009.bam')
 print(myWell)
 # An S4 object of class WellSample
 #
 # Well_ID :  W384
 # Well_Number :  384
 # BAMfilename :  extdata/P009.bam
 #
}
\seealso{
\code{\link{initsWellSamples}}
}

