% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasing_functions.R
\name{getMutationsOfLFR}
\alias{getMutationsOfLFR}
\title{Retrieve the  mutations on an LFR}
\usage{
getMutationsOfLFR(LFR_info, Mutations_df, calltype = NULL)
}
\description{
This function computes the list of mutations present on a  Long Fragment Reads either on their Variant Allele or their reference Allele
}

